requirements.txt
scikit-learn
matplotlib
pandas 
numpy
xgboost
streamlit
google-genai
pydantic
dotenv
ipykernel
kagglehub